backend/requirements.txt
flask
flask-cors
transformers
torch
opencv-python
pillow
numpy
safetensors
